codeJunior/requirements.txt
asgiref==3.7.2
bleach==6.1.0
Django==4.2.4
django-summernote==0.8.20.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
Pillow==9.5.0
psycopg2-binary==2.9.7
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
webencodings==0.5.1
